YCR: init images module
This module will be responsible for managing images (storing into DB, calculating blurhash, saving to S3).
