+++ src/ya-comiste-rust/server/Cargo.toml
@@ -10,11 +10,13 @@ publish = false
 [dependencies]
 arangors = "0.4.5"
 async-trait = "0.1.42"
+blurhash-wasm = { git = "https://github.com/fpapado/blurhash-rust-wasm", branch = "master" }
 data-encoding = "2.3.1"
 dataloader = "0.14.0"
 deadpool = { version = "0.7.0", features = ["managed"] }
 env_logger = "0.8.2"
 futures = "0.3.12"
+image = "0.23.12"
 jsonwebtoken = "7.2.0"
 juniper = "0.15.2"
 lazy_static = "1.4.0"
@@ -23,6 +25,8 @@ num_cpus = "1.13.0"
 rand = "0.8.2"
 reqwest = { version = "0.10.10", features = ["json", "blocking"] }
 ring = "0.16.5"
+rusoto_core = "0.45.0"
+rusoto_s3 = "0.45.0"
 serde = "1.0.119"
 serde_json = "1.0.61"
 signedsource = { path = "../signedsource" }
