+++ src/ya-comiste-rust/server/src/sdui/sdui_card_component.rs
@@ -1,5 +1,5 @@
 use crate::graphql_context::Context;
-use crate::sdui::resolve_cloudfront_url::resolve_cloudfront_url;
+use crate::images::resolve_cloudfront_url;
 use crate::sdui::sdui_content::SDUIContent;
 use juniper::{FieldError, FieldResult};
